chore(observations): extract calculateRecursiveMetricsForRunItems helper from getObservationsGroupedByTraceId
